review: Move back strategy to server
